+++ MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/service/TaskService.java
@@ -27,11 +27,11 @@ public class TaskService {
         .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
   }
 
-  public List<Task> findByAssignedTo(Long userId) {
+  public List<Task> findByAssignedTo(int userId) {
     return taskRepository.findByAssignedTo(userId);
   }
 
-  public List<Task> findBySprintAndUser(int sprintId, Long userId) {
+  public List<Task> findBySprintAndUser(int sprintId, int userId) {
     return findByAssignedTo(userId)
         .stream()
         .filter(task -> task.getID_Sprint() == sprintId)
